Virker ikke btw

+++ scripts/Road.h
@@ -1,16 +1,13 @@
 
 
 
-struct road
-{
-    int speedLimit;
+struct road{
+
+    double speedLimit;
     int lanes;
     int startX;
     int startY;
     int endX;
     int endY;
-
-
-
 
 };
